+++ src/Pim/Bundle/BaseConnectorBundle/DependencyInjection/PimBaseConnectorExtension.php
@@ -23,7 +23,6 @@ class PimBaseConnectorExtension extends Extension
     {
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('archiving.yml');
-        $loader->load('converters.yml');
         $loader->load('processors.yml');
         $loader->load('readers.yml');
         $loader->load('steps.yml');
