+++ src/Pim/Bundle/BaseConnectorBundle/Resources/config/processors.yml
@@ -13,6 +13,10 @@ parameters:
     pim_base_connector.processor.denormalization.csv_variant_group.class:   Pim\Bundle\BaseConnectorBundle\Processor\Denormalization\VariantGroupProcessor
     pim_base_connector.processor.denormalization.attribute_option.class:    Pim\Bundle\BaseConnectorBundle\Processor\Denormalization\AttributeOptionProcessor
 
+    pim_base_connector.denormalization.array_converter.flat.product.class:                Pim\Bundle\BaseConnectorBundle\Processor\Denormalization\ArrayConverter\Flat\ProductToStandardConverter
+    pim_base_connector.denormalization.array_converter.flat.attribute_option.class:       Pim\Bundle\BaseConnectorBundle\Processor\Denormalization\ArrayConverter\Flat\AttributeOptionToStandardConverter
+    pim_base_connector.denormalization.array_converter.structured.attribute_option.class: Pim\Bundle\BaseConnectorBundle\Processor\Denormalization\ArrayConverter\Structured\AttributeOptionToStandardConverter
+
     pim_base_connector.processor.product_to_flat_array.class:           Pim\Bundle\BaseConnectorBundle\Processor\ProductToFlatArrayProcessor
     pim_base_connector.processor.normalization.csv_variant_group.class: Pim\Bundle\BaseConnectorBundle\Processor\Normalization\VariantGroupProcessor
     pim_base_connector.processor.dummy.class:                           Pim\Bundle\BaseConnectorBundle\Processor\DummyProcessor
@@ -155,7 +159,7 @@ services:
     pim_base_connector.processor.denormalization.attribute_option.flat:
         class: %pim_base_connector.processor.denormalization.attribute_option.class%
         arguments:
-            - '@pim_base_connector.converter.flat.attribute_option'
+            - '@pim_base_connector.denormalization.array_converter.flat.attribute_option'
             - '@pim_catalog.repository.attribute_option'
             - '@pim_catalog.updater.attribute_option'
             - '@pim_validator'
@@ -165,7 +169,7 @@ services:
     pim_base_connector.processor.denormalization.attribute_option.structured:
         class: %pim_base_connector.processor.denormalization.attribute_option.class%
         arguments:
-            - '@pim_base_connector.converter.structured.attribute_option'
+            - '@pim_base_connector.denormalization.array_converter.structured.attribute_option'
             - '@pim_catalog.repository.attribute_option'
             - '@pim_catalog.updater.attribute_option'
             - '@pim_validator'
@@ -175,7 +179,7 @@ services:
     pim_base_connector.processor.denormalization.product.flat:
         class: %pim_base_connector.processor.denormalization.product.class%
         arguments:
-            - '@pim_base_connector.converter.flat.product'
+            - '@pim_base_connector.denormalization.array_converter.flat.product'
             - '@pim_catalog.repository.product'
             - '@pim_catalog.builder.product'
             - '@pim_catalog.updater.product'
@@ -194,6 +198,21 @@ services:
             - %pim_catalog.entity.product.class%
             - 'csv'
 
+    # Denormalization / Converters
+    pim_base_connector.denormalization.array_converter.flat.product:
+        class: %pim_base_connector.denormalization.array_converter.flat.product.class%
+        arguments:
+            - '@pim_transform.builder.field_name'
+
+    pim_base_connector.denormalization.array_converter.flat.attribute_option:
+        class: %pim_base_connector.denormalization.array_converter.flat.attribute_option.class%
+        arguments:
+            - '@pim_catalog.repository.locale'
+
+    pim_base_connector.denormalization.array_converter.structured.attribute_option:
+        class: %pim_base_connector.denormalization.array_converter.structured.attribute_option.class%
+
+    # Normalization
     pim_base_connector.processor.normalization.csv_group:
         class: %pim_base_connector.processor.csv_serializer.homogeneous.class%
         parent: pim_base_connector.processor.csv_serializer
